+++ src/main/scala/com/atomist/rug/functions/travis/DisableRepoFunction.scala
@@ -15,7 +15,7 @@ class DisableRepoFunction
     * @param owner GitHub owner, i.e., user or organization, of the repo to enable
     * @param repo name of the repo to disable
     * @param org Travis CI ".com" or ".org" endpoint
-    * @param token GitHub token with "repo" scope for `owner`/`repo`
+    * @param githubToken GitHub token with proper scopes for Travis CI
     * @return Rug Function Response
     */
   @RugFunction(name = "travis-disable-repo", description = "Disables Travis CI builds for a GitHub repository",
@@ -24,10 +24,10 @@ class DisableRepoFunction
                @Parameter(name = "owner") owner: String,
                @Parameter(name = "repo") repo: String,
                @Parameter(name = "org") org: String,
-               @Secret(name = "github_token", path = "github://user_token?scopes=repo") token: String
+               @Secret(name = "githubToken", path = TravisFunction.githubTokenPath) githubToken: String
              ): FunctionResponse = {
     val disableTravis = false
-    RepoHook(travisEndpoints).tryRepoHook(disableTravis, owner, repo, token, org)
+    RepoHook(travisEndpoints).tryRepoHook(disableTravis, owner, repo, githubToken, org)
   }
 
 }